AdjacencyMatrix: double -> int

+++ src/adjacencymatrix.hpp
@@ -9,18 +9,18 @@ class AdjacencyMatrix
 public:
     AdjacencyMatrix(int size);
 
-    const std::vector<double> &operator[](int column) const;
-    std::vector<double> &operator[](int column);
+    const std::vector<int> &operator[](int column) const;
+    std::vector<int> &operator[](int column);
 
     int getSize() const;
 
     bool isValid() const;
 
-    void generate(double min, double max);
+    void generate(int min, int max);
     void clear();
 
 private:
-    std::vector<std::vector<double>> m_data;
+    std::vector<std::vector<int>> m_data;
 
     static std::default_random_engine m_randomEngine;
